requirements.txt
aiogram>=3.0
python-dotenv
deep-translator